worked on routine requests, sending them works okay-ish-ly too between five versions

+++ net/proto/net_proto_api.h
@@ -7,7 +7,7 @@ namespace net_proto{
 	namespace request{
 		void add_id(id_t_ id);
 		void add_id(std::vector<id_t_> id);
-		void add_id_linked_list(id_t_ id, uint32_t length);
+		void add_id_linked_list(id_t_ id, int64_t length);
 		void del_id(id_t_ id);
 		void del_id(std::vector<id_t_> id);
 		void add_fast_routine_type(std::string type);
